% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDtrace.R
\name{convert_indices_to_edge}
\alias{convert_indices_to_edge}
\title{SPD-trace: Efficient Learning of Differential Networks}
\usage{
convert_indices_to_edge(upper_tri_indices, p, node_labels)
}
\description{
SPD-trace: Efficient Learning of Differential Networks
}
