% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{kendall_tau_matrix_to_pearson_correlation_matrix}
\alias{kendall_tau_matrix_to_pearson_correlation_matrix}
\title{Convert Kendall's Tau Matrix to Pearson Correlation Matrix}
\usage{
kendall_tau_matrix_to_pearson_correlation_matrix(kendall_tau_matrix)
}
\arguments{
\item{kendall_tau_matrix}{Matrix of Kendall's tau correlation coefficients}
}
\value{
Matrix of Pearson correlation coefficients
}
\description{
Converts a matrix of Kendall's tau correlation coefficients to Pearson correlation 
coefficients using the transformation: r = sin(π/2 * τ).
}
\examples{
# Create example Kendall's tau matrix
tau_matrix <- matrix(c(1, 0.5, 0.3, 0.5, 1, 0.4, 0.3, 0.4, 1), 3, 3)
pearson_matrix <- kendall_tau_matrix_to_pearson_correlation_matrix(tau_matrix)
print(pearson_matrix)

}
