% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{positive_semi_definite_maker}
\alias{positive_semi_definite_maker}
\title{Make Matrix Positive Semi-Definite}
\usage{
positive_semi_definite_maker(A)
}
\arguments{
\item{A}{Symmetric matrix to be made positive semi-definite}
}
\value{
Positive semi-definite matrix
}
\description{
Ensures a symmetric matrix is positive semi-definite by adjusting diagonal elements
if necessary. This is useful for correlation matrices that may have negative eigenvalues.
}
\examples{
# Create a matrix that might not be positive semi-definite
A <- matrix(c(1, 0.9, 0.9, 1), 2, 2)
A_psd <- positive_semi_definite_maker(A)
eigen(A_psd)$values  # Should all be >= 0

}
