% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDtrace-package.R
\docType{package}
\name{SPDtrace-package}
\alias{SPDtrace-package}
\title{SPDtrace Package}
\description{
Efficient Learning of Differential Networks in Multi-Source Non-Paranormal Graphical Models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mojtabanikahd/SPDtrace}
  \item Report bugs at \url{https://github.com/mojtabanikahd/SPDtrace/issues}
}

}
\author{
\strong{Maintainer}: Mojtaba Nikahd \email{nikahd@ce.sharif.ir}

Authors:
\itemize{
  \item Seyed Abolfazl Motahari
}

}
\keyword{internal}
