% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDtrace.R
\name{print.SPDtrace_result}
\alias{print.SPDtrace_result}
\title{Print method for SPDtrace results}
\usage{
\method{print}{SPDtrace_result}(x, ...)
}
\arguments{
\item{x}{Object of class SPDtrace_result}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for SPDtrace results
}
