% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDtrace.R
\name{summary.SPDtrace_result}
\alias{summary.SPDtrace_result}
\title{Summary method for SPDtrace results}
\usage{
\method{summary}{SPDtrace_result}(object, ...)
}
\arguments{
\item{object}{Object of class SPDtrace_result}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary method for SPDtrace results
}
